Updates
Fix all tracks CRUD APIs
Fix all Courses CRUD APIs

+++ controller/course_controller.js
@@ -1,16 +1,48 @@
-const Course = require('../mongoose.models/course');
+const {Course} = require('../mongoose.models/course');
 const Track = require('../mongoose.models/track');
 const asyncWrapper = require('../middleware/asyncWrapper');
 const Member = require('../mongoose.models/member');
 const jwt = require('../middleware/jwt');
+const { selectFields } = require('express-validator/lib/field-selection');
 
 // ================== COURSE CRUD ==================
 
 // Create a new course
 const createCourse = asyncWrapper(async (req, res) => {
-    const { name, description, tracks } = req.body;
+    const {email} = req.decoded;
+    console.log(email); 
+    const member = await Member.findOne({ email });
+    console.log("validating member");
+    if(!member) {
+        return res.status(404).json({ success: false, message: 'Member not found' });
+    }
+    if(member.role != "head"){
+        return res.status(403).json({ success: false, message: 'Only heads can create courses' });
+    }
+    console.log("validating body"); 
+    const { name, description, trackId, admins , committee } = req.body;
+    if(!name || !description || !trackId || !admins || !committee) {
+        return res.status(400).json({ success: false, message: 'Please provide all required fields' });
+    }
+    console.log("validating track and admins");
+    const track = await Track.findById(trackId);
+    if(!track) {
+        return res.status(404).json({ success: false, message: 'Track not found' });
+    }
+    if(member.committee != track.committee){
+        return res.status(403).json({ success: false, message: 'You can only create courses for your committee tracks' });
+    }
 
-    const course = new Course({ name, description, tracks });
+    for(const adminId of admins){
+        const admin = await Member.findById(adminId);
+        if(!admin) {
+            return res.status(404).json({ success: false, message: `Admin with ID ${adminId} not found` });
+        }
+    }
+
+    const course = new Course({ name, description, tracks: [track._id], admins , committee });
+    track.courses.push(course._id);
+    await track.save();
     await course.save();
 
     res.status(201).json({ success: true, data: course });
@@ -18,7 +50,7 @@ const createCourse = asyncWrapper(async (req, res) => {
 
 // Get all courses
 const getAllCourses = asyncWrapper(async (req, res) => {
-    const courses = await Course.find().populate('tracks').populate('tasks.assignees');
+    const courses = await Course.find().populate('admins', 'name email  committee avatar');
     res.status(200).json({ success: true, data: courses });
 });
 
@@ -33,24 +65,54 @@ const getCourseById = asyncWrapper(async (req, res) => {
 
 // Update a course
 const updateCourse = asyncWrapper(async (req, res) => {
-    const { name, description, tracks } = req.body;
-    const course = await Course.findByIdAndUpdate(
-        req.params.id,
-        { name, description, tracks },
-        { new: true }
-    );
-    if (!course) {
+    const { email } = req.decoded;
+    console.log(email);
+    const member = await Member.findOne({ email });
+    if(!member) {
+        return res.status(404).json({ success: false, message: 'Member not found' });
+    }
+    if(member.role != "head"){
+        return res.status(403).json({ success: false, message: 'Only heads can update courses' });
+    }
+
+    const courseToUpdate = await Course.findById(req.params.id);
+    if(!courseToUpdate) {
         return res.status(404).json({ success: false, message: 'Course not found' });
     }
+    if(courseToUpdate.committee != member.committee){
+        return res.status(403).json({ success: false, message: 'You can only update courses in your committee' });
+    }
+
+    const { name, description } = req.body;
+    if(!name || !description) {
+        return res.status(400).json({ success: false, message: 'Please provide all required fields' });
+    }
+    const course = await Course.findByIdAndUpdate(
+        req.params.id,
+        { name, description },
+        { new: true }
+    );
     res.status(200).json({ success: true, data: course });
 });
 
 // Delete a course
 const deleteCourse = asyncWrapper(async (req, res) => {
-    const course = await Course.findByIdAndDelete(req.params.id);
+    const { email } = req.decoded;
+    const member = await Member.findOne({ email});
+    if(!member) {
+        return res.status(404).json({ success: false, message: 'Member not found' });
+    }
+    if(member.role != "head"){
+        return res.status(403).json({ success: false, message: 'Only heads can delete courses' });
+    }
+    const course = await Course.findOne({ _id: req.params.id });
     if (!course) {
         return res.status(404).json({ success: false, message: 'Course not found' });
     }
+    if(member.committee != course.committee){
+        return res.status(403).json({ success: false, message: 'You can only delete courses in your committee' });
+    }
+    await Course.findByIdAndDelete(req.params.id);
     res.status(200).json({ success: true, message: 'Course deleted successfully' });
 });
 
@@ -58,13 +120,31 @@ const deleteCourse = asyncWrapper(async (req, res) => {
 
 // Add track to course
 const addTrackToCourse = asyncWrapper(async (req, res) => {
-    const { trackId } = req.body;
-    const course = await Course.findById(req.params.id);
-    if (!course) {
-        return res.status(404).json({ success: false, message: 'Course not found' });
+    const { email } = req.decoded;
+    const member = await Member.findOne({ email});
+    if(!member) {
+        return res.status(404).json({ success: false, message: 'Member not found' });
     }
-    if (!course.tracks.includes(trackId)) {
-        course.tracks.push(trackId);
+    if(member.role != "head"){
+        return res.status(403).json({ success: false, message: 'Only heads can assign courses' });
+    }
+
+
+        var track = await Track.findById(req.params.trackId);
+        var course = await Course.findById(req.params.courseId);
+        if(member.committee != track.committee){
+            return res.status(403).json({ success: false, message: 'You can only assign to tracks in your committee' });
+        }
+        if (!track) {
+            return res.status(404).json({ success: false, message: 'Track not found' });
+        }
+        if (!course) {
+            return res.status(404).json({ success: false, message: 'Course not found' });
+        }
+    if (!course.tracks.includes(req.params.trackId)) {
+        course.tracks.push(req.params.trackId);
+        track.courses.push(course._id);
+        await track.save(); 
         await course.save();
     }
     res.status(200).json({ success: true, data: course });
@@ -72,15 +152,30 @@ const addTrackToCourse = asyncWrapper(async (req, res) => {
 
 // Remove track from course
 const removeTrackFromCourse = asyncWrapper(async (req, res) => {
-    const { trackId } = req.body;
-    const course = await Course.findByIdAndUpdate(
-        req.params.id,
-        { $pull: { tracks: trackId } },
-        { new: true }
-    );
+    const { email } = req.decoded;
+    const member = await Member.findOne({ email});
+    if(!member) {
+        return res.status(404).json({ success: false, message: 'Member not found' });
+    }
+    if(member.role != "head"){
+        return res.status(403).json({ success: false, message: 'Only heads can assign courses' });
+    }
+
+    var track = await Track.findById(req.params.trackId);
+    var course = await Course.findById(req.params.courseId);
+    if(member.committee != track.committee){
+        return res.status(403).json({ success: false, message: 'You can only assign to tracks in your committee' });
+    }
+    if (!track) {
+        return res.status(404).json({ success: false, message: 'Track not found' });
+    }
     if (!course) {
         return res.status(404).json({ success: false, message: 'Course not found' });
     }
+    course.tracks = course.tracks.filter(track => track != req.params.trackId);
+    track.courses = track.courses.filter(course => course != req.params.courseId);
+    await track.save();
+    await course.save();
     res.status(200).json({ success: true, data: course });
 });
